Lots of docs, bugfixes, and some refactoring

+++ src/fileops.c
@@ -25,6 +25,7 @@ fileop_progress()
 	return &m_progress;
 }
 
+/* Chmod a file, and if it's a directory, chmod all of its contents as well */
 int
 chmod_file(char *name, mode_t mode)
 {
@@ -37,6 +38,7 @@ chmod_file(char *name, mode_t mode)
 	return 0;
 }
 
+/* Copy a file, and if it's a directory, all of its contents as well */
 int
 copy_file(char *src, char *dest)
 {
@@ -59,22 +61,25 @@ copy_file(char *src, char *dest)
 	}
 
 	m_progress.fname = NULL;
-
-	return 0;
+	return copy_status;
 }
 
+/* Delete a file, and if it's a directory, all of its contents as well */
 int
 delete_file(char *name)
 {
+	int delete_status;
+
 	m_progress.obj_count = enumerate_dir(name);
 	m_progress.obj_done = 0;
 
-	s_delete_file(name);
+	delete_status = s_delete_file(name);
 
 	m_progress.fname = NULL;
-	return 0;
+	return delete_status;
 }
 
+/* Symlink a file, and only that file, even if it is a directory */
 int
 link_file(char *src, char *dest)
 {
@@ -84,20 +89,36 @@ link_file(char *src, char *dest)
 	return 0;
 }
 
+/* Move something by trying to use rename() since it's faster and atomic by 
+ * definition. If this does not work, resort to copying src to dest and deleting
+ * the source. If anything fails during the copy, bail out, since we can't be 
+ * sure the files have made it safely to their new destination */
 int
 move_file(char *src, char *dest)
 {
 	int retval;
-	if ((retval = copy_file(src, dest)) < 0) {
-		return retval;
-	}
-	if ((retval = delete_file(src)) < 0) {
-		return retval;
+
+	retval = 0;
+	if (rename(src, dest)) {    /* Try to rename atomically */
+		if (errno == EXDEV) {   /* We're moving across filesystems */
+			if ((retval = copy_file(src, dest)) < 0) {
+				return retval;
+			}
+			if ((retval = delete_file(src)) < 0) {
+				return retval;
+			}
+        } else {
+            return errno;
+        }
 	}
 	return retval;
 }
 
 /* Static functions {{{ */
+/* Enumerate a directory contents to guesstimate how long an operation will
+ * approximately take. Note than a finer-grained estimate would slow down the
+ * operation significatly, because it wouldn't be carried out in a single
+ * sendfile() call like it is now */
 unsigned
 enumerate_dir(char *path)
 {
@@ -105,18 +126,18 @@ enumerate_dir(char *path)
 	struct dirent *ep;
 	struct stat st;
 	char *subpath;
-	unsigned ret;
+	unsigned count;
 
-	ret = 1;
+	count = 1;
 	if ((dp = opendir(path))) {
 		while ((ep = readdir(dp))) {
 			if (!is_dot_or_dotdot(ep->d_name)) {
-				ret++;
+				count++;
 				subpath = join_path(path, ep->d_name);
 				lstat(subpath, &st);
 
 				if (S_ISDIR(st.st_mode)) {
-					ret += enumerate_dir(subpath);
+					count += enumerate_dir(subpath);
 				}
 
 				free(subpath);
@@ -125,11 +146,11 @@ enumerate_dir(char *path)
 		closedir(dp);
 	}
 
-	return ret;
+	return count;
 }
 
 /* Recursive chmodding of a directory and its children. Possible TODO: add an
- * option to just chmod the top level */
+ * option to just chmod the top level file */
 int
 s_chmod_file(char *name, mode_t mode)
 {
@@ -156,77 +177,76 @@ s_chmod_file(char *name, mode_t mode)
 	return 0;
 }
 
-/* Recursive copying TODO: make this iterative */
+/* Recursive copying backend. Possible TODO: make this iterative */
 int
 s_copy_file(char *src, char *dest)
 {
-	int in_fd, out_fd;
-	struct stat st;
-	DIR *dp;
-	struct dirent *ep;
 	char *subpath_src, *subpath_dest;
+	int in_fd, out_fd;
 	int retval;
-
-	retval = -1;
+	DIR *dp;
+	struct stat st;
+	struct dirent *ep;
 
 	m_progress.fname = dest;
 
-	if ((in_fd = open(src, O_RDONLY)) >= 0) {
-		fstat(in_fd, &st);
+	retval = -1;
+	if ((in_fd = open(src, O_RDONLY)) < 0) {
+		return retval;
+	}
 
-		/* Recursive copy if it's a directory */
-		if (S_ISDIR(st.st_mode)) {
-			close(in_fd);
-			mkdir(dest, st.st_mode);
-			m_progress.obj_done++;
-
-			if ((dp = opendir(src))) {
-				while ((ep = readdir(dp))) {
-					if (!is_dot_or_dotdot(ep->d_name)) {
-						subpath_src = join_path(src, ep->d_name);
-						subpath_dest = join_path(dest, ep->d_name);
-
-						retval = s_copy_file(subpath_src, subpath_dest);
-						/* Clean up the mess if the copy failed */
-						switch (retval) {
-						case ENOMEM:    /* 4 intentional fallthroughs */
-						case EINVAL:
-						case EOVERFLOW:
-						case EIO:
-							delete_file(subpath_dest);
-							break;
-						default:
-							break;
-						}
-						queue_master_update();
-
-						free(subpath_src);
-						free(subpath_dest);
+	if (!(dp = fdopendir(in_fd))) {     /* Try to open in_fd as a directory */
+		retval = errno;
+		if (errno == ENOTDIR) {         /* Src is a file: fstat and copy it */
+			if (!fstat(in_fd, &st)) {
+				if ((out_fd = open(dest, O_WRONLY|O_CREAT, &st.st_mode)) >= 0) {
+					if (sendfile(out_fd, in_fd, NULL, st.st_size) < 0) {
+						retval = errno;
+					} else {
+						retval = 0;
 					}
+					close(out_fd);
 				}
-				closedir(dp);
-			}
-		} else {
-			if ((out_fd = open(dest, O_WRONLY|O_CREAT, &st.st_mode)) >= 0) {
-				if (sendfile(out_fd, in_fd, NULL, st.st_size) < 0) {
-					retval = errno;
-				} else {
-					retval = 0;
-				}
-
-				m_progress.obj_done++;
-				close(out_fd);
 			}
 		}
 		close(in_fd);
+	} else {                            /* Src is a directory: recursive copy */
+		mkdir(dest, st.st_mode);        /* """copy""" the directory */
+		m_progress.obj_done++;
+
+		/* Scan the directory, and self-call on each node contained inside */
+		while ((ep = readdir(dp))) {
+			if (!is_dot_or_dotdot(ep->d_name)) {    /* Skip . and .. */
+				subpath_src = join_path(src, ep->d_name);
+				subpath_dest = join_path(dest, ep->d_name);
+				retval = s_copy_file(subpath_src, subpath_dest);
+				free(subpath_src);
+				free(subpath_dest);
+
+				/* Clean up the mess if the copy failed (read: delete the file 
+				 * if something bad happened during the copy */
+				switch (retval) {
+				case ENOMEM:            /* 4 intentional fallthroughs */
+				case EINVAL:
+				case EOVERFLOW:
+				case EIO:
+					delete_file(subpath_dest);
+					break;
+				default:
+					break;
+				}
+
+			}
+		}
+		closedir(dp);                   /* No need to close(in_fd) */
 	}
 
-	m_progress.fname = NULL;
-
+    m_progress.obj_done++;
+	queue_master_update();              /* Request an UI update */
 	return retval;
 }
 
-/* Recursive deletion TODO: make this iterative */
+/* Recursive deletion. Possible TODO: make this iterative */
 int
 s_delete_file(char *name)
 {
@@ -239,7 +259,7 @@ s_delete_file(char *name)
 	m_progress.fname = name;
 
 	/* Delete a directory's contents before attempting to delete the directory
-	 * itself */
+	 * itself to prevent an ENOTEMPTY */
 	if(S_ISDIR(st.st_mode) && (dp = opendir(name))) {
 		while ((ep = readdir(dp))) {
 			if (!is_dot_or_dotdot(ep->d_name)) {
@@ -251,6 +271,7 @@ s_delete_file(char *name)
 		closedir(dp);
 	}
 
+	/* Delete the file/folder itself */
 	if (remove(name) < 0) {
 		return errno;
 	}
